refactor: refactoring export methods, quote patterns and small structure fixes
